Add colored scrollbar. Begin About Me. Temp tweaks
Made a few last-second tweaks in order to experiment with different looks for each section. As a result, this is in a rather messy state! I am thinking of making the particle background in a fixed position across the entire page, and introducing some transparency to various elements in each section in order for those particles to bleed through.

I also colored the scrollbar to share the site's red accent theme. I then began an About Me section.