Make Contact less cramped, About less boxy
Just various small changes in the Contact section to give it a little more room to breathe. The About Me section felt a little boxy, so the border and background around the main paragraph has been removed. The top margin on the Traits section was also reduced, as it was off in the middle of nowhere and looked like its own section!