Recolor favicon to blue (new primary color)
The favicon is the only part of the website that does not adapt when I change the "color-primary" variable. As I feel rather comfortable with the new blue color scheme I chose for my portfolio, I went ahead and re-colored the favicon in Photoshop to use align with this scheme.