Re-order sections, add tech logos
After hearing from a few sources that Projects should be front-and-center, I decided to move the Projects section to appear just below the fold. I'm personally on the fence with this decision, but for now it will stay.

Also added img tags to each "Skill" and linked them with their appropriate SVG.